@@ -27,64 +27,66 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.6",
-    "@angular/cdk": "5.2.3",
-    "@angular/common": "5.2.6",
-    "@angular/compiler": "5.2.6",
-    "@angular/core": "5.2.6",
-    "@angular/flex-layout": "2.0.0-beta.12",
-    "@angular/forms": "5.2.6",
-    "@angular/http": "5.2.6",
-    "@angular/material": "5.2.3",
-    "@angular/platform-browser": "5.2.6",
-    "@angular/platform-browser-dynamic": "5.2.6",
-    "@angular/router": "5.2.6",
-    "@nationallibraryofnorway/ngx-mime": "../dist/",
+    "@angular/animations": "6.0.0-rc.5",
+    "@angular/cdk": "6.0.0-rc.5",
+    "@angular/common": "6.0.0-rc.5",
+    "@angular/compiler": "6.0.0-rc.5",
+    "@angular/core": "6.0.0-rc.5",
+    "@angular/flex-layout": "6.0.0-beta.15",
+    "@angular/forms": "6.0.0-rc.5",
+    "@angular/http": "6.0.0-rc.5",
+    "@angular/material": "6.0.0-rc.5",
+    "@angular/platform-browser": "6.0.0-rc.5",
+    "@angular/platform-browser-dynamic": "6.0.0-rc.5",
+    "@angular/router": "6.0.0-rc.5",
+    "@nationallibraryofnorway/ngx-mime": "../dist/ngx-mime",
     "classlist.js": "^1.1.20150312",
-    "core-js": "^2.4.1",
-    "d3": "^4.10.0",
+    "core-js": "2.5.5",
+    "d3": "4.1.0",
     "openseadragon": "2.3.1",
-    "rxjs": "5.5.2",
-    "web-animations-js": "^2.2.5",
-    "zone.js": "0.8.14"
+    "rxjs": "6.0.0-uncanny-rc.7",
+    "web-animations-js": "2.3.1",
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.1",
-    "@angular/compiler-cli": "5.2.6",
-    "@angular/language-service": "5.2.6",
-    "@types/chai": "4.0.4",
+    "@angular/cli": "6.0.0-rc.5",
+    "@angular/compiler-cli": "6.0.0-rc.5",
+    "@angular-devkit/build-angular": "0.5.6",
+    "@angular/language-service": "6.0.0-rc.5",
+    "@types/chai": "4.1.2",
     "@types/chai-as-promised": "7.1.0",
     "@types/cucumber": "2.0.4",
-    "@types/fs-extra": "4.0.3",
-    "@types/jasmine": "~2.5.53",
-    "@types/node": "~6.0.60",
-    "bluebird": "^3.5.0",
-    "body-parser": "^1.18.2",
+    "@types/fs-extra": "5.0.2",
+    "@types/jasmine": "2.8.6",
+    "@types/node": "9.6.5",
+    "@types/es6-promise": "3.3.0",
+    "bluebird": "3.5.1",
+    "body-parser": "1.18.2",
     "chai": "4.1.2",
     "chai-as-promised": "7.1.1",
-    "chromedriver": "2.33.2",
-    "codelyzer": "4.0.1",
-    "concurrently": "^3.4.0",
-    "cucumber": "3.1.0",
-    "express": "^4.16.2",
-    "fs-extra": "4.0.1",
-    "geckodriver": "1.9.0",
-    "jasmine-core": "2.8.0",
-    "jasmine-spec-reporter": "~4.1.0",
-    "jsonfile": "3.0.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "mountebank": "^1.11.0",
-    "protractor": "5.2.0",
-    "protractor-cucumber-framework": "4.1.1",
-    "protractor-multiple-cucumber-html-reporter-plugin": "^1.2.3",
-    "rimraf": "^2.5.4",
-    "ts-node": "3.3.0",
-    "tslint": "5.7.0",
-    "typescript": "2.5.3"
+    "chromedriver": "2.37.0",
+    "codelyzer": "4.2.1",
+    "concurrently": "3.5.1",
+    "cucumber": "4.2.1",
+    "express": "4.16.3",
+    "fs-extra": "5.0.0",
+    "geckodriver": "1.11.0",
+    "jasmine-core": "3.1.0",
+    "jasmine-spec-reporter": "4.2.1",
+    "jsonfile": "4.0.0",
+    "karma": "2.0.0",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-cli": "1.0.1",
+    "karma-coverage-istanbul-reporter": "1.4.2",
+    "karma-jasmine": "1.1.1",
+    "karma-jasmine-html-reporter": "1.0.0",
+    "mountebank": "1.14.0",
+    "protractor": "5.3.1",
+    "protractor-cucumber-framework": "5.0.0",
+    "protractor-multiple-cucumber-html-reporter-plugin": "1.7.0",
+    "rimraf": "2.6.2",
+    "ts-node": "5.0.1",
+    "tslint": "5.9.1",
+    "typescript": "2.7.2"
   }
 }
