@@ -22,7 +22,7 @@
     "precommit": "pretty-quick --staged",
     "clean": "rimraf out-tsc dist \"src/lib/**/*.js\" \"src/lib/**/*.map\"",
     "prebuild": "npm run clean",
-    "build": "node build.js",
+    "build": "ng-packagr -p package.json",
     "build-demo": "tsc -p src/demo/",
     "build-demo:watch": "tsc -p src/demo/ -w",
     "build-scss": "node-sass src/lib -o src/lib --output-style expanded",
@@ -48,35 +48,35 @@
     "release": "sh scripts/release/release.sh"
   },
   "peerDependencies": {
-    "@angular/cdk": ">=5.0.0 <6.0.0",
-    "@angular/core": ">=5.0.0 <6.0.0",
-    "@angular/flex-layout": ">=2.0.0 <3.0.0 || >=2.0.0-beta.12 <3.0.0",
-    "@angular/material": ">=5.0.0 <6.0.0",
+    "@angular/cdk": ">=6.0.0 <7.0.0 || >=6.0.0-rc.5 < 7.0.0",
+    "@angular/core": ">=6.0.0 <7.0.0 || >=6.0.0-rc.5 < 7.0.0",
+    "@angular/flex-layout": ">=6.0.0 <7.0.0 || >=6.0.0-beta.15 <7.0.0",
+    "@angular/material": ">=6.0.0 <7.0.0 || >=6.0.0-rc.5 < 7.0.0",
     "d3": ">=4.1.0 <5.0.0",
     "openseadragon": ">=2.3.1 <3.0.0"
   },
   "devDependencies": {
-    "@angular/animations": "5.2.6",
-    "@angular/cdk": "5.2.3",
-    "@angular/common": "5.2.6",
-    "@angular/compiler": "5.2.6",
-    "@angular/compiler-cli": "5.2.6",
-    "@angular/core": "5.2.6",
-    "@angular/flex-layout": "2.0.0-beta.12",
-    "@angular/forms": "5.2.6",
-    "@angular/http": "5.2.6",
-    "@angular/language-service": "5.2.6",
-    "@angular/material": "5.2.3",
-    "@angular/platform-browser": "5.2.6",
-    "@angular/platform-browser-dynamic": "5.2.6",
-    "@angular/router": "5.2.6",
+    "@angular/animations": "6.0.0-rc.5",
+    "@angular/cdk": "6.0.0-rc.5",
+    "@angular/common": "6.0.0-rc.5",
+    "@angular/compiler": "6.0.0-rc.5",
+    "@angular/compiler-cli": "6.0.0-rc.5",
+    "@angular/core": "6.0.0-rc.5",
+    "@angular/flex-layout": "6.0.0-beta.15",
+    "@angular/forms": "6.0.0-rc.5",
+    "@angular/http": "6.0.0-rc.5",
+    "@angular/language-service": "6.0.0-rc.5",
+    "@angular/material": "6.0.0-rc.5",
+    "@angular/platform-browser": "6.0.0-rc.5",
+    "@angular/platform-browser-dynamic": "6.0.0-rc.5",
+    "@angular/router": "6.0.0-rc.5",
     "@types/d3": "^4.10.0",
     "@types/jasmine": "2.5.54",
     "@types/node": "^6.0.46",
     "camelcase": "^4.0.0",
-    "codelyzer": "4.0.1",
+    "codelyzer": "4.2.1",
     "concurrently": "3.4.0",
-    "core-js": "^2.4.1",
+    "core-js": "2.5.5",
     "d3": "^4.10.0",
     "execa": "^0.6.3",
     "glob": "^7.1.1",
@@ -92,25 +92,20 @@
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-remap-coverage": "^0.1.4",
     "lite-server": "2.2.2",
+    "ng-packagr": "2.4.2",
     "node-sass": "^4.5.2",
     "openseadragon": "2.3.1",
-    "prettier": "1.10.2",
+    "prettier": "1.12.1",
     "pretty-quick": "^1.4.1",
     "rimraf": "^2.6.1",
-    "rollup": "0.56.3",
-    "rollup-plugin-commonjs": "8.3.0",
-    "rollup-plugin-node-resolve": "3.0.3",
-    "rollup-plugin-sourcemaps": "0.4.2",
-    "rollup-plugin-uglify": "3.0.0",
-    "rxjs": "5.5.2",
-    "standard-version": "^4.0.0",
-    "systemjs": "0.20.19",
-    "tslib": "1.7.1",
+    "rxjs": "6.0.0-uncanny-rc.7",
+    "standard-version": "4.3.0",
+    "systemjs": "0.21.2",
+    "tslib": "1.9.0",
     "tslint": "5.9.1",
-    "tslint-config-prettier": "1.9.0",
-    "typescript": "2.5.3",
-    "uglify-es": "^3.0.18",
+    "tslint-config-prettier": "1.10.0",
+    "typescript": "2.7.2",
     "web-animations-js": "^2.3.1",
-    "zone.js": "0.8.18"
+    "zone.js": "0.8.26"
   }
 }
