@@ -2,9 +2,8 @@ import { async, ComponentFixture, TestBed, fakeAsync, flush } from '@angular/cor
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { By } from '@angular/platform-browser';
 import { MatDialogRef, ShowOnDirtyErrorStateMatcher, ErrorStateMatcher } from '@angular/material';
-import { distinctUntilChanged } from 'rxjs/operators/distinctUntilChanged';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
-import { Observable } from 'rxjs/Observable';
+import { Observable, BehaviorSubject } from 'rxjs';
+import { distinctUntilChanged } from 'rxjs/operators';
 
 import { SharedModule } from '../shared/shared.module';
 import { CanvasGroupDialogComponent } from './canvas-group-dialog.component';
@@ -25,24 +24,22 @@ describe('PageDialogComponent', () => {
   let intl: MimeViewerIntl;
   let canvasService: CanvasServiceStub;
 
-  beforeEach(
-    async(() => {
-      TestBed.configureTestingModule({
-        imports: [NoopAnimationsModule, SharedModule],
-        declarations: [CanvasGroupDialogComponent],
-        providers: [
-          ViewerService,
-          ClickService,
-          ModeService,
-          ViewerLayoutService,
-          MimeViewerIntl,
-          { provide: IiifContentSearchService, useClass: IiifContentSearchServiceStub },
-          { provide: MatDialogRef, useClass: MatDialogRefStub },
-          { provide: CanvasService, useClass: CanvasServiceStub }
-        ]
-      }).compileComponents();
-    })
-  );
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [NoopAnimationsModule, SharedModule],
+      declarations: [CanvasGroupDialogComponent],
+      providers: [
+        ViewerService,
+        ClickService,
+        ModeService,
+        ViewerLayoutService,
+        MimeViewerIntl,
+        { provide: IiifContentSearchService, useClass: IiifContentSearchServiceStub },
+        { provide: MatDialogRef, useClass: MatDialogRefStub },
+        { provide: CanvasService, useClass: CanvasServiceStub }
+      ]
+    }).compileComponents();
+  }));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(CanvasGroupDialogComponent);
