@@ -4,7 +4,7 @@ import { By } from '@angular/platform-browser';
 import { async, ComponentFixture, inject, TestBed } from '@angular/core/testing';
 import { ObservableMedia } from '@angular/flex-layout';
 import { MatDialogRef } from '@angular/material';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 import { SharedModule } from '../../shared/shared.module';
 import { MimeViewerIntl } from '../../core/intl/viewer-intl';
@@ -26,24 +26,22 @@ describe('TocComponent', () => {
   let fixture: ComponentFixture<TocComponent>;
   let iiifManifestService: IiifManifestServiceStub;
 
-  beforeEach(
-    async(() => {
-      TestBed.configureTestingModule({
-        imports: [SharedModule, HttpClientModule],
-        declarations: [TocComponent],
-        providers: [
-          ClickService,
-          CanvasService,
-          ModeService,
-          MimeViewerIntl,
-          { provide: MatDialogRef, useClass: MatDialogRefStub },
-          { provide: ObservableMedia, useClass: MediaServiceStub },
-          { provide: IiifManifestService, useClass: IiifManifestServiceStub },
-          { provide: ViewerService, useClass: ViewerServiceStub }
-        ]
-      }).compileComponents();
-    })
-  );
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [SharedModule, HttpClientModule],
+      declarations: [TocComponent],
+      providers: [
+        ClickService,
+        CanvasService,
+        ModeService,
+        MimeViewerIntl,
+        { provide: MatDialogRef, useClass: MatDialogRefStub },
+        { provide: ObservableMedia, useClass: MediaServiceStub },
+        { provide: IiifManifestService, useClass: IiifManifestServiceStub },
+        { provide: ViewerService, useClass: ViewerServiceStub }
+      ]
+    }).compileComponents();
+  }));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(TocComponent);
